requirements.txt
Django==1.11.4
djangorestframework==3.6.4
Markdown==2.6.9
pytz==2017.2
